# api/v1/model.go
// Package v1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.4 DO NOT EDIT.
package v1

// Account defines model for Account.
type Account struct {
	// Id The ID
	Id ID `binding:"required,max=36" json:"id"`

	// Name Account name
	Name string `binding:"required,max=255" json:"name"`
}

// Error defines model for Error.
type Error struct {
	// Code Error code
	Code int8 `binding:"required,gte=0,lte=127" json:"code"`

	// Message Error message
	Message string `binding:"required,max=4000" json:"message"`
}

// ID The ID
type ID = string

// NewAccount defines model for NewAccount.
type NewAccount struct {
	// Name Account name
	Name string `binding:"required,max=255" json:"name"`
}

// AddAccountJSONRequestBody defines body for AddAccount for application/json ContentType.
type AddAccountJSONRequestBody = NewAccount