# api/v1/openapi.yaml
openapi: "3.0.3"
info:
  version: 1.0.0
  title: Financier API
  description: OpenAPI specification of the Financier service
  contact:
    name: Financier developer
    email: olegshishkin@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /api/v1
paths:
  /accounts:
    get:
      tags:
        - Accounts
      description: Returns all accounts.
      operationId: getAllAccounts
      parameters: [ ]
      responses:
        '200':
          description: account response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Accounts
      description: Creates new account.
      operationId: addAccount
      requestBody:
        description: Account to add.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewAccount'
      responses:
        '201':
          description: account creation response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /accounts/{id}:
    get:
      tags:
        - Accounts
      description: Returns an account by its ID.
      operationId: findAccountByID
      parameters:
        - name: id
          in: path
          description: ID of account to fetch
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: account response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          description: no account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    ID:
      description: The ID
      type: string
      nullable: false
      minLength: 1
      maxLength: 36
      example: 123
      x-oapi-codegen-extra-tags:
        binding: "required,max=36"
    NewAccount:
      type: object
      properties:
        name:
          description: Account name
          type: string
          nullable: false
          minLength: 1
          maxLength: 255
          example: wallet
          x-oapi-codegen-extra-tags:
            binding: "required,max=255"
      required:
        - name
      additionalProperties: false
    Account:
      allOf:
        - type: object
          properties:
            id:
              $ref: '#/components/schemas/ID'
          required:
            - id
        - $ref: '#/components/schemas/NewAccount'
    Error:
      type: object
      properties:
        code:
          description: Error code
          type: integer
          format: int8
          minimum: 0
          maximum: 127
          example: 1
          x-oapi-codegen-extra-tags:
            binding: "required,gte=0,lte=127"
        message:
          description: Error message
          type: string
          nullable: false
          minLength: 1
          maxLength: 4000
          example: Unexpected error
          x-oapi-codegen-extra-tags:
            binding: "required,max=4000"
      required:
        - code
        - message
      additionalProperties: false